+++ src/core_services/apple_events/mod.rs
@@ -2,8 +2,10 @@ mod desc;
 mod event_class;
 mod event_id;
 mod return_id;
+mod transaction_id;
 
 pub use desc::*;
 pub use event_class::*;
 pub use event_id::*;
 pub use return_id::*;
+pub use transaction_id::*;