+++ CHANGELOG.md
@@ -118,6 +118,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 
     - `AEReturnID` newtype struct over `i16`.
 
+    - `AETransactionID` newtype struct over `i32`.
+
 - Created `core_animation` module for
   [Core Animation](https://developer.apple.com/documentation/quartzcore)
   framework.